requirements.txt
pandas>=1.3.0
openpyxl>=3.0.9
python-dotenv>=0.19.0
